chore(nlu): module lifecyles are indenpendant of core nlu logic
